+++ EmployeeApp/SystemEngineer.cs
@@ -7,7 +7,7 @@ namespace EmployeeApp
         public string WorkPlace { get; set; }
         public SystemEngineer(int employeeId, string firstName, string lastName, 
                             string email, string phone, Gender gender, Address address,
-                            int hourlyPrice, int systemsDesigned, double bonus, double totalHours)
+                            double hourlyPrice, int systemsDesigned, double bonus, double totalHours)
         : base(employeeId, firstName, lastName, email, phone, gender, address, hourlyPrice, systemsDesigned, bonus, totalHours){ }
         public override string getJobName()
         {
